refactor elastic package
