+++ common/elastic/client.go
@@ -3,63 +3,10 @@ package elastic
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/MurmurationsNetwork/MurmurationsServices/common/logger"
-	"github.com/cenkalti/backoff"
 	"github.com/olivere/elastic"
 )
-
-const docType = "_doc"
-
-var (
-	Client esClientInterface = &esClient{}
-)
-
-func NewClient(url string) error {
-	var client *elastic.Client
-
-	op := func() error {
-		log := logger.GetLogger()
-
-		var err error
-		client, err = elastic.NewClient(
-			elastic.SetURL(url),
-			elastic.SetHealthcheckInterval(10*time.Second),
-			elastic.SetErrorLog(log),
-			elastic.SetInfoLog(log),
-		)
-		if err != nil {
-			return err
-		}
-
-		return nil
-	}
-	notify := func(err error, time time.Duration) {
-		logger.Info(fmt.Sprintf("trying to re-connect ElasticSearch %s \n", err))
-	}
-	b := backoff.NewExponentialBackOff()
-	b.MaxElapsedTime = 2 * time.Minute
-
-	err := backoff.RetryNotify(op, b, notify)
-	if err != nil {
-		return err
-	}
-
-	Client.setClient(client)
-
-	return nil
-}
-
-type esClientInterface interface {
-	setClient(*elastic.Client)
-
-	CreateMappings([]Index) error
-	Index(string, interface{}) (*elastic.IndexResponse, error)
-	IndexWithID(string, string, interface{}) (*elastic.IndexResponse, error)
-	Search(string, elastic.Query) (*elastic.SearchResult, error)
-	Delete(string, string) error
-}
 
 type esClient struct {
 	client *elastic.Client
