+++ common/elastic/constant.go
@@ -1,0 +1,3 @@
+package elastic
+
+const docType = "_doc"
